Beta 1.2

+++ src/main/java/ch/hekates/languify/language/LangLoader.java
@@ -1,11 +1,9 @@
 package ch.hekates.languify.language;
 
 import ch.hekates.languify.Languify;
-import org.bukkit.plugin.Plugin;
 
 import java.io.*;
 import java.nio.file.Files;
-import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.util.Collections;
 import java.util.zip.ZipEntry;
@@ -13,18 +11,23 @@ import java.util.zip.ZipFile;
 
 public class LangLoader {
     /**
-     * This method loads the set language by checking if the file already exists. If not, it clones the file out of the plugin jar file into the language folder.
-     *
-     * @param language The language file name without the file ending .json.
-     * @see Languify#setFileDirectory(String)
+     * Sets the specified language.
+     * @param language
+     * @deprecated Use {@link Languify#setLanguage} instead.
      */
-    private static final Plugin plugin = Languify.getPlugin();
-
+    @Deprecated
     public static void loadLanguage(String language) {
         Languify.setLanguage(language);
     }
 
-    public static void saveLanguages(String pluginName, String version) {
+    /**
+     * Saves the language files from the jar into the plugins folder.
+     * @param pluginName The name of the plugin (as specified in the plugin.yml). Used to locate the plugin folder.
+     * @see Languify#setOverWriteFiles(boolean)
+     * @deprecated Probably useless for most use cases. Use {@link #saveLanguages(String)} and append the version to the plugin name.
+     */
+    @Deprecated
+    public static void saveLanguages(String pluginName, @Deprecated String version) {
         File targetDirectory = new File("plugins/" + pluginName);
         String jarFilePath = "plugins/" + pluginName + version + ".jar";
 
@@ -37,7 +40,42 @@ public class LangLoader {
                     targetFile.getParentFile().mkdirs();
 
                     try (InputStream inputStream = jarFile.getInputStream(entry)) {
-                        Files.copy(inputStream, targetFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                        if (Languify.getOverWriteFiles()) {
+                            Files.copy(inputStream, targetFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                        } else {
+                            Files.copy(inputStream, targetFile.toPath());
+                        }
+                    }
+                }
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    /**
+     * Saves the language files from the jar into the plugins folder.
+     * @param pluginName The name of the plugin (as specified in the plugin.yml). Used to locate the plugin folder.
+     * @see Languify#setOverWriteFiles(boolean)
+     */
+    public static void saveLanguages(String pluginName) {
+        File targetDirectory = new File("plugins/" + pluginName);
+        String jarFilePath = "plugins/" + pluginName + ".jar";
+
+        try (ZipFile jarFile = new ZipFile(jarFilePath)) {
+            for (ZipEntry entry : Collections.list(jarFile.entries())) {
+                if (entry.getName().startsWith("lang/") && !entry.isDirectory()) {
+                    String targetFilePath = targetDirectory.getPath() + "/" + entry.getName();
+
+                    File targetFile = new File(targetFilePath);
+                    targetFile.getParentFile().mkdirs();
+
+                    try (InputStream inputStream = jarFile.getInputStream(entry)) {
+                        if (Languify.getOverWriteFiles()) {
+                            Files.copy(inputStream, targetFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                        } else {
+                            Files.copy(inputStream, targetFile.toPath());
+                        }
                     }
                 }
             }
